# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentação da Api-dicionário destinada a auxiliar na individualização de CSAs do projeto Agromart",
    "version" : "1.0.0",
    "title" : "Api-Dicionario",
    "license" : {
      "name" : "MIT",
      "url" : "https://github.com/mucahitnezir/express-starter/blob/master/LICENSE"
    }
  },
 
  "tags" : [  {
    "name" : "csa"
  } ],
  "paths" : {
   
    "/csa" : {
      "get" : {
        "summary" : "Lista todas as CSAs cadastradas",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "type" : "integer",
          "required" : false,
          "description" : "Page"
        }, {
          "in" : "query",
          "name" : "perPage",
          "type" : "integer",
          "required" : false,
          "description" : "Per page"
        } ],
        "tags" : [ "csa" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/ListaDeCSA"
            }
          }
        }
      },
      "post" : {
        "summary" : "Cadastra uma CSA",
        "parameters" : [ {
          "name" : "body",
          "in" : "body",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "urlBase","nomeCSA","responsavelCSA","emailCSA" ],
            "properties" : {
              "urlBase" : {
                "type" : "string"
              },"nomeCSA" : {
                "type" : "string"
              },"responsavelCSA" : {
                "type" : "string"
              },"emailCSA" : {
                "type" : "string"
              }
            }
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/CSA"
            }
          }
         
        },
        "tags" : [ "csa" ],
        "description" : "Cadastra uma CSA"
      }
    },
    "/csa/{id}" : {
      "get" : {
        "summary" : "Resgata informações da CSA por ID",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "type" : "integer",
          "required" : true,
          "description" : "Id da CSA"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/CSA"
            }
          }
        },
        "tags" : [ "csa" ],
        "description" : "Resgata informações da CSA por ID"
      },
      "delete" : {
        "summary" : "Deletar uma CSA por ID",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "type" : "integer",
          "required" : true,
          "description" : "Id da CSA"
        } ],
        "responses" : {
          "204" : {
            "description" : "OK"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "tags" : [ "csa" ],
        "description" : "Deletar uma CSA por ID "
      }
    }
  },
  "definitions" : {
   
    "CSA" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "nomeCSA" : {
          "type" : "string"
        },
        "urlBase" : {
          "type" : "string"
        },
        "responsavelCSA" : {
            "type" : "string"
          },
        "createdAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updatedAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "deletedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
      }
    },
    "ListaDeCSA" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer"
          },
          "perPage" : {
            "type" : "integer"
          },
          "totalPage" : {
            "type" : "integer"
          },
          "count" : {
            "type" : "integer"
          },
          "rows" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/CSA"
            }
          }
        }
      }
   
  }
}